refactor: rename history.tsx to router.tsx
