+++ src/Initial.tsx
@@ -2,7 +2,7 @@ import { observer } from "mobx-react-lite";
 import { Flex, Typography } from "antd";
 import { useEffect } from "react";
 import { locales, modules } from "./modules";
-import { initRouter } from "./history";
+import { initRouter } from "./router";
 import { gstate } from "./global";
 import { Indicator } from "./components/Indicator";
 import { avifCheck } from "./engines/support";
